LPS-172038 portal-search-web: pull up testOneTerm
https://issues.liferay.com/browse/LPS-172038O
