+++ modules/apps/portal-search/portal-search-web/src/test/java/com/liferay/portal/search/web/internal/facet/display/context/AssetTagsSearchFacetDisplayContextTest.java
@@ -74,41 +74,6 @@ public class AssetTagsSearchFacetDisplayContextTest
 	}
 
 	@Test
-	public void testOneTerm() throws Exception {
-		String term = RandomTestUtil.randomString();
-		int frequency = RandomTestUtil.randomInt();
-
-		setUpTermCollectors(
-			facetCollector,
-			Collections.singletonList(createTermCollector(term, frequency)));
-
-		String facetParam = StringPool.BLANK;
-
-		FacetDisplayContext facetDisplayContext = createFacetDisplayContext(
-			facetParam);
-
-		List<BucketDisplayContext> bucketDisplayContexts =
-			facetDisplayContext.getBucketDisplayContexts();
-
-		Assert.assertEquals(
-			bucketDisplayContexts.toString(), 1, bucketDisplayContexts.size());
-
-		BucketDisplayContext bucketDisplayContext = bucketDisplayContexts.get(
-			0);
-
-		Assert.assertEquals(term, bucketDisplayContext.getBucketText());
-		Assert.assertEquals(term, bucketDisplayContext.getFilterValue());
-		Assert.assertEquals(frequency, bucketDisplayContext.getFrequency());
-		Assert.assertFalse(bucketDisplayContext.isSelected());
-		Assert.assertTrue(bucketDisplayContext.isFrequencyVisible());
-
-		Assert.assertEquals(
-			facetParam, facetDisplayContext.getParameterValue());
-		Assert.assertTrue(facetDisplayContext.isNothingSelected());
-		Assert.assertFalse(facetDisplayContext.isRenderNothing());
-	}
-
-	@Test
 	public void testOneTermWithPreviousSelection() throws Exception {
 		String term = RandomTestUtil.randomString();
 		int frequency = RandomTestUtil.randomInt();
