+++ modules/apps/portal-search/portal-search-web/src/test/java/com/liferay/portal/search/web/internal/facet/display/context/BaseCategoriesSearchFacetDisplayContextTestCase.java
@@ -146,6 +146,7 @@ public abstract class BaseCategoriesSearchFacetDisplayContextTestCase
 		_excludedGroupId = 0;
 	}
 
+	@Override
 	@Test
 	public void testOneTerm() throws Exception {
 		long assetCategoryId = RandomTestUtil.randomLong();
